+++ TimeHACK.Main/WinClassic/TestApp.cs
@@ -1,0 +1,13 @@
+﻿using System.Windows.Forms;
+
+namespace TimeHACK.WinClassicForms
+{
+    public partial class TestApp : UserControl
+    {
+        public TestApp()
+        {
+            InitializeComponent();
+        }
+    }
+}
+
