+++ TimeHACK.Main/WinClassic/WinClassicCMD.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace HistacomCS
+{
+    public partial class WinClassicCommandPrompt : Form
+    {
+        public WinClassicCommandPrompt()
+        {
+            InitializeComponent();
+        }
+    }
+}
