+++ TimeHACK.Main/WinClassic/WinClassicTemplate.cs
@@ -1,0 +1,12 @@
+﻿using System.Windows.Forms;
+
+namespace TimeHACK
+{
+    public partial class WinClassicTemplate : Form
+    {
+        public WinClassicTemplate()
+        {
+            InitializeComponent();
+        }
+    }
+}
