--- TimeHACK.Main/WinClassicForms/TestApp.cs
@@ -1,13 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace TimeHACK.WinClassicForms
-{
-    public partial class TestApp : UserControl
-    {
-        public TestApp()
-        {
-            InitializeComponent();
-        }
-    }
-}
-
