--- TimeHACK.Main/WinClassicForms/WinClassicCMD.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace HistacomCS
-{
-    public partial class WinClassicCommandPrompt : Form
-    {
-        public WinClassicCommandPrompt()
-        {
-            InitializeComponent();
-        }
-    }
-}
