--- TimeHACK.Main/WinClassicForms/WinClassicTemplate.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace TimeHACK
-{
-    public partial class WinClassicTemplate : Form
-    {
-        public WinClassicTemplate()
-        {
-            InitializeComponent();
-        }
-    }
-}
