+++ TimeHACK.Main/bin/Release/TimeHACK.application
@@ -14,7 +14,7 @@
           <dsig:Transform Algorithm="urn:schemas-microsoft-com:HashTransforms.Identity" />
         </dsig:Transforms>
         <dsig:DigestMethod Algorithm="http://www.w3.org/2000/09/xmldsig#sha256" />
-        <dsig:DigestValue>a/Rg0oqaFV3brH5B97iOCDXc0yawkl9S/2cVtgXXS8M=</dsig:DigestValue>
+        <dsig:DigestValue>owcvndBWt3qGkDndsMykYlr5kB7zH3NDbzGfSUEMI4Q=</dsig:DigestValue>
       </hash>
     </dependentAssembly>
   </dependency>
